fix(bidi): ensure consistent interface between service and directive
Currently the `bidi` module has a `Directionality` service and a `Dir` directive. The directive re-provides itself as the service, however it doesn't have the same interface and behavior can cause inconsistencies depending on where it's injected. E.g. the directive has a `dir` property whereas the service doesn't. Also the directive will emit when it's value is re-assigned, but the service won't.

These changes bring the two closer together by moving the functionality into the service and having the directive extend it.