fix(select): have form control disabled state take precedence over disabled binding
`MatSelect` can be disabled either through the `disabled` property or via a `ControlValueAccessor`, but currently it's inconsistent if the two values conflict with each other since we take the value of the last one that was set. These changes make it so that the value from the CVA has precedence. This is consistent with how we handle the disabled state in `MatInput`.

Fixes #19437.